- final(?) commit

+++ src/main/java/de/lemo/dms/connectors/moodle/ExtractAndMapMoodle.java
@@ -250,10 +250,10 @@ public class ExtractAndMapMoodle extends ExtractAndMap {
 			for(Object[] obj : tmpl)
 			{
 				ForumPostsLMS p = new ForumPostsLMS();
-				p.setId(((BigInteger) obj[0]).longValue());
+				p.setId(((Integer) obj[0]).longValue());
 				p.setUserid(((String) obj[1]));
-				p.setCreated(((BigInteger) obj[2]).longValue());
-				p.setModified(((BigInteger) obj[3]).longValue());
+				p.setCreated(((Integer) obj[2]).longValue());
+				p.setModified(((Integer) obj[3]).longValue());
 				p.setSubject((String) obj[4]);
 				p.setMessage((String) obj[5]);
 				this.forumPostsLms.add(p);
